+++ sbt_common/messaging/src/test/scala/uk/ac/wellcome/messaging/test/fixtures/SQS.scala
@@ -2,9 +2,10 @@ package uk.ac.wellcome.messaging.test.fixtures
 
 import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration
 import com.amazonaws.services.sqs._
-import com.amazonaws.services.sqs.model.PurgeQueueRequest
+import com.amazonaws.services.sqs.model.{PurgeQueueRequest, ReceiveMessageRequest}
 import com.amazonaws.auth.BasicAWSCredentials
 import com.amazonaws.auth.AWSStaticCredentialsProvider
+import org.scalatest.Matchers
 import uk.ac.wellcome.messaging.sqs.SQSMessage
 import uk.ac.wellcome.test.fixtures._
 
@@ -20,7 +21,7 @@ object SQS {
 
 }
 
-trait SQS {
+trait SQS extends Matchers {
 
   import SQS._
 
@@ -126,6 +127,40 @@ trait SQS {
         body = """{ "foo": "bar"}""",
         timestamp = "timestamp"
       )
+  }
+
+  def assertQueueEmpty(queue: Queue) = {
+    // The visibility timeout is set to 1 second for test queues.
+    // Wait for slightly longer than that to make sure that messages
+    // that fail processing become visible again before asserting.
+    Thread.sleep(1500)
+
+    val messages = sqsClient
+      .receiveMessage(
+        new ReceiveMessageRequest(queue.url)
+          .withMaxNumberOfMessages(1)
+      )
+      .getMessages
+      .toList
+
+    messages shouldBe empty
+  }
+
+  def assertQueueNotEmpty(queue: Queue) = {
+    // The visibility timeout is set to 1 second for test queues.
+    // Wait for slightly longer than that to make sure that messages
+    // that fail processing become visible again before asserting.
+    Thread.sleep(1500)
+
+    val messages = sqsClient
+      .receiveMessage(
+        new ReceiveMessageRequest(queue.url)
+          .withMaxNumberOfMessages(1)
+      )
+      .getMessages
+      .toList
+
+    messages should not be empty
   }
 
 }
